fix: fixed hard string
