+++ apps/web/utils/creatorProfile.ts
@@ -1,3 +1,29 @@
+export type Passwords = {
+  current: string;
+  next: string;
+  confirm: string;
+};
+
+export type ProfileForm = {
+  firstName: string;
+  lastName: string;
+  company: string;
+  phone: string;
+  cvr: string;
+  address: string;
+  city: string;
+  postal: string;
+  reg: string;
+  account: string;
+  email: string;
+};
+
+export type PasswordState = {
+  current: string;
+  next: string;
+  confirm: string;
+};
+
 export type PaymentKeys = "reg" | "account";
 export type CompanyKeys =
   | "company"
@@ -6,12 +32,6 @@ export type CompanyKeys =
   | "address"
   | "city"
   | "postal";
-
-export type Passwords = {
-  current: string;
-  next: string;
-  confirm: string;
-};
 
 export const FORM_KEYS = {
   firstName: "firstName",
